Backend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
mpmath==1.3.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.12.0
